requirements.txt
requests
openai
anthropic
fastapi
uvicorn[standard]
pyairtable
python-dotenv
